[Feature #160011175] Refactor tests
